backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
firebase-admin==7.1.0
cryptography==45.0.7
python-multipart==0.0.20
qrcode[pil]==8.2
reportlab==4.4.4
Pillow==11.3.0
numpy==1.26.4
opencv-python==4.11.0.86
pypdfium2==5.7.0
torch==2.2.2
torchvision==0.17.2
